Simplify handling of JSON values being sent out.
With this change, any time you provide either a dict or a list as a
return value from a handler, that object is by default returned with a
JSON content type.

A string defaults to an HTML content type.

If you want to return other content types, there are currently two ways
to do this:
* Explicitly statically serve files, in which case the filename
  extension for many common types will be recognized and served as the
  content type; otherwise
* Return a tuple from your handler function. The first value is returned
  as the content type (so you can set it yourself); the second value is
  the response body.