By default, treat string output as html.
Before this change, we only considered `bytes` values to be html. My
original thinking was that what we send over the wire is bytes. But in
practice, I seem to be finding it convenient to not worry about the
string->bytes conversion, and to let shotglass handle it for me.